Move readme parser out of the build
