+++ source/features/rgh-feature-descriptions.tsx
@@ -5,8 +5,8 @@ import CopyIcon from 'octicons-plain-react/Copy';
 import InfoIcon from 'octicons-plain-react/Info';
 
 import features from '../feature-manager.js';
-import {featuresMeta} from '../../readme.md';
-import optionsStorage, {getNewFeatureName, isFeatureDisabled} from '../options-storage.js';
+import optionsStorage, {isFeatureDisabled} from '../options-storage.js';
+import {featuresMeta, getNewFeatureName} from '../feature-data.js';
 import observe from '../helpers/selector-observer.js';
 import {brokenFeatures} from '../helpers/hotfix.js';
 import createRghIssueLink from '../helpers/rgh-issue-link.js';
