+++ source/globals.d.ts
@@ -10,7 +10,7 @@ type FeatureID = string & {feature: true};
 interface FeatureMeta {
 	id: FeatureID;
 	description: string;
-	screenshot?: string;
+	screenshot: string | null; // eslint-disable-line @typescript-eslint/ban-types -- We use `null` in the JSON file
 }
 
 // These types are unnecessarily loose
@@ -30,11 +30,6 @@ interface Window {
 }
 
 declare module 'size-plugin';
-
-declare module '*.md' { // It should be just for readme.md, but 🤷‍♂️
-	export const importedFeatures: FeatureID[];
-	export const featuresMeta: FeatureMeta[];
-}
 
 declare module '*.gql' {
 	export = string;
