+++ source/options-storage.ts
@@ -1,7 +1,6 @@
 import OptionsSyncPerDomain from 'webext-options-sync-per-domain';
 
-import renamedFeatures from './feature-renames.json';
-import {importedFeatures} from '../readme.md';
+import {importedFeatures, renamedFeatures} from './feature-data.js';
 
 export type RGHOptions = typeof defaults;
 
@@ -17,12 +16,6 @@ export function isFeatureDisabled(options: RGHOptions, id: string): boolean {
 	// Must check if it's specifically `false`: It could be undefined if not yet in the readme or if misread from the entry point #6606
 	// eslint-disable-next-line @typescript-eslint/no-unnecessary-boolean-literal-compare
 	return options[`feature:${id}`] === false;
-}
-
-export function getNewFeatureName(possibleFeatureName: string): FeatureID | undefined {
-	// @ts-expect-error Useless "no index type" error as usual
-	const newFeatureName = renamedFeatures[possibleFeatureName] as FeatureID ?? possibleFeatureName;
-	return importedFeatures.includes(newFeatureName) ? newFeatureName : undefined;
 }
 
 const migrations = [
